# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LMS V1 API
  version: 1.0.0
  description: Canonical single-owner LMS contract. NestJS is the authorization authority.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security: []
paths:
...
  /owner/courses:
    get:
      tags: [Owner courses]
      operationId: listOwnerCourses
      summary: List all courses for owner operations
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Owner course page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoursePage" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Owner courses]
      operationId: createCourse
      summary: Create a draft course
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CourseWriteRequest" }
      responses:
        "201":
          description: Draft course created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /owner/courses/{courseId}:
    get:
      tags: [Owner courses]
      operationId: getOwnerCourse
      summary: Get a course and its complete curriculum for owner editing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Owner curriculum detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CourseWithCurriculum" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Owner courses]
      operationId: updateCourse
      summary: Update course metadata
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CourseWriteRequest" }
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /owner/courses/{courseId}/publish:
    post:
      tags: [Owner courses]
      operationId: publishCourse
      summary: Publish a valid course
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Course published
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /owner/courses/{courseId}/archive:
    post:
      tags: [Owner courses]
      operationId: archiveCourse
      summary: Archive a course
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Course archived
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /owner/courses/{courseId}/sections:
    post:
      tags: [Owner courses]
      operationId: createSection
      summary: Create a course section
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SectionWriteRequest" }
      responses:
        "201":
          description: Section created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Section" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /owner/courses/{courseId}/sections/order:
    put:
      tags: [Owner courses]
      operationId: reorderSections
      summary: Atomically reorder every course section
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderedIdsRequest" }
      responses:
        "204": { description: Sections reordered atomically }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /owner/sections/{sectionId}:
    patch:
      tags: [Owner courses]
      operationId: updateSection
      summary: Update a course section
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SectionWriteRequest" }
      responses:
        "200":
          description: Section updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Section" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /owner/sections/{sectionId}/lessons:
    post:
      tags: [Owner courses]
      operationId: createLesson
      summary: Create a section lesson
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SectionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LessonWriteRequest" }
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lesson" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /owner/sections/{sectionId}/lessons/order:
    put:
      tags: [Owner courses]
      operationId: reorderLessons
      summary: Atomically reorder every section lesson
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderedIdsRequest" }
      responses:
        "204": { description: Lessons reordered atomically }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /owner/lessons/{lessonId}:
    patch:
      tags: [Owner courses]
      operationId: updateLesson
      summary: Update a lesson
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LessonId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LessonWriteRequest" }
      responses:
        "200":
          description: Lesson updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lesson" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /owner/courses/{courseId}/enrollments:
    get:
      tags: [Owner enrollments]
      operationId: listCourseEnrollments
      summary: List enrollment records for one course
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Enrollment page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnrollmentPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Owner enrollments]
      operationId: grantEnrollment
      summary: Grant course access to a student
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnrollmentWriteRequest" }
      responses:
        "201":
          description: Enrollment granted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Enrollment" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /owner/enrollments/{enrollmentId}:
    patch:
      tags: [Owner enrollments]
      operationId: updateEnrollment
      summary: Update an enrollment access window or suspension
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/EnrollmentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnrollmentWriteRequest" }
      responses:
        "200":
          description: Enrollment updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Enrollment" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /owner/students:
    get:
      tags: [Owner users]
      operationId: searchStudents
      summary: Search students using a least-data operational response
      security: [{ bearerAuth: [] }]
      parameters:
        - name: query
          in: query
          schema: { type: string, maxLength: 200 }
      responses:
        "200":
          description: Student page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /owner/enrollments/{enrollmentId}/revoke:
    post:
      tags: [Owner enrollments]
      operationId: revokeEnrollment
      summary: Permanently revoke an enrollment grant
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/EnrollmentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Enrollment revoked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Enrollment" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: lms_refresh
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 512 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    CourseId: { name: courseId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
    SectionId: { name: sectionId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
    LessonId: { name: lessonId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
...
    EnrollmentId: { name: enrollmentId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
...
  responses:
    ValidationFailed:
      description: Request validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Forbidden:
      description: Authorization or account policy denied access
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found or concealed by authorization
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: State or uniqueness conflict
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    Error:
      type: object
      additionalProperties: false
      required: [statusCode, code, message, requestId]
      properties:
        statusCode: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { $ref: "#/components/schemas/Uuid" }
...
    Course:
      type: object
      additionalProperties: false
      required: [id, title, slug, description, status, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        thumbnailUrl: { type: [string, "null"], format: uri }
        status: { type: string, enum: [DRAFT, PUBLISHED, ARCHIVED] }
        publishedAt: { type: [string, "null"], format: date-time }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
    CoursePage:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Course" } }
        nextCursor: { type: [string, "null"] }
    CourseWithCurriculum:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          required: [sections]
          properties:
            sections: { type: array, items: { $ref: "#/components/schemas/Section" } }
    CourseWriteRequest:
      type: object
      additionalProperties: false
      required: [title, slug, description]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$", maxLength: 200 }
        description: { type: string, minLength: 1, maxLength: 20000 }
    Section:
      type: object
      required: [id, courseId, title, sortOrder, lessons]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        courseId: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        description: { type: [string, "null"] }
        sortOrder: { type: integer, minimum: 0 }
        lessons: { type: array, items: { $ref: "#/components/schemas/Lesson" } }
    SectionWriteRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"], maxLength: 10000 }
    Lesson:
      type: object
      required: [id, sectionId, title, type, sortOrder]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        sectionId: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        description: { type: [string, "null"] }
        type: { type: string, enum: [VIDEO, TEXT] }
        textContent: { type: [string, "null"] }
        sortOrder: { type: integer, minimum: 0 }
        progress: { oneOf: [{ $ref: "#/components/schemas/LessonProgress" }, { type: "null" }] }
        resources:
          type: array
          items: { $ref: "#/components/schemas/LessonResource" }
    LessonResource:
      type: object
      additionalProperties: false
      required: [id, lessonId, title, filename, mimeType, sizeBytes, createdAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        lessonId: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: string, pattern: "^[1-9][0-9]*$" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
...
    LessonWriteRequest:
      type: object
      additionalProperties: false
      required: [title, type]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"], maxLength: 10000 }
        type: { type: string, enum: [VIDEO, TEXT] }
        textContent: { type: [string, "null"], maxLength: 100000 }
    OrderedIdsRequest:
      type: object
      additionalProperties: false
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: "#/components/schemas/Uuid" }
    EnrollmentWriteRequest:
      type: object
      additionalProperties: false
      required: [userId, startsAt]
      properties:
        userId: { $ref: "#/components/schemas/Uuid" }
        startsAt: { $ref: "#/components/schemas/Timestamp" }
        expiresAt: { type: [string, "null"], format: date-time }
        status: { type: string, enum: [ACTIVE, SUSPENDED] }
    Enrollment:
      type: object
      required: [id, userId, courseId, status, startsAt, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        userId: { $ref: "#/components/schemas/Uuid" }
        courseId: { $ref: "#/components/schemas/Uuid" }
        status: { type: string, enum: [ACTIVE, SUSPENDED, REVOKED, EXPIRED] }
        startsAt: { $ref: "#/components/schemas/Timestamp" }
        expiresAt: { type: [string, "null"], format: date-time }
        completedAt: { type: [string, "null"], format: date-time }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
    EnrollmentPage:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Enrollment" } }
        nextCursor: { type: [string, "null"] }
    StudentSummary:
      type: object
      additionalProperties: false
      required: [id, email, firstName, lastName, status, createdAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        status: { type: string, enum: [PENDING_VERIFICATION, ACTIVE, SUSPENDED, DISABLED] }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
    StudentPage:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/StudentSummary" } }
        nextCursor: { type: [string, "null"] }
...
    LessonProgress:
      type: object
      required: [lessonId, lastPositionSeconds, maximumPositionSeconds, watchedSeconds, percentage]
      properties:
        lessonId: { $ref: "#/components/schemas/Uuid" }
        lastPositionSeconds: { type: integer, minimum: 0 }
        maximumPositionSeconds: { type: integer, minimum: 0 }
        watchedSeconds: { type: integer, minimum: 0 }
        percentage: { type: number, minimum: 0, maximum: 100 }
        completedAt: { type: [string, "null"], format: date-time }
...